Updated Design of Tools
Improved Design by unifying margins across tools and also unified the button sizes for consistency.
