--- src/routes/(tools)/line-sorter/+page.svelte
@@ -1,137 +1,0 @@
-
-<script lang="ts">
-
-	import { Button } from 'flowbite-svelte';
-	import jsPDF from 'jspdf';
-	import Intro from '$lib/Intro.svelte';
-  
-	export let data;
-  
-	let inputString = '';
-	let sortedString = '';
-	let sortingOption = "Select"; // Default value
-  
-	function handleInputChange(event) {
-	  inputString = event.target.value;
-	  sortText();
-	}
-  
-	function sortText() {
-	  switch (sortingOption) {
-		case "alphabetical":
-		  // Split the input into lines and sort them alphabetically
-		  const linesAlphabetical = inputString.split('\n').filter(line => line.trim() !== '').sort((a, b) => a.localeCompare(b));
-		  sortedString = linesAlphabetical.join('\n');
-		  break;
-		case "reverse":
-		  // Split the input into lines and reverse their order
-		  const linesReverse = inputString.split('\n').filter(line => line.trim() !== '').reverse();
-		  sortedString = linesReverse.join('\n');
-		  break;
-		case "random":
-		  // Split the input into lines and shuffle them randomly
-		  const linesRandom = inputString.split('\n').filter(line => line.trim() !== '').sort(() => 0.5 - Math.random());
-		  sortedString = linesRandom.join('\n');
-		  break;
-		case "length":
-		  // Split the input into lines and sort them by line length
-		  const linesByLength = inputString.split('\n').filter(line => line.trim() !== '').sort((a, b) => a.length - b.length);
-		  sortedString = linesByLength.join('\n');
-		  break;
-		default:
-		  sortedString = inputString;
-	  }
-	}
-  
-	function copyText() {
-	  if (sortedString.length > 0) {
-		var textarea = document.createElement("textarea");
-		textarea.value = sortedString;
-		document.body.appendChild(textarea);
-		textarea.select();
-		document.execCommand("copy");
-		document.body.removeChild(textarea);
-	  }
-	}
-  
-	function downloadText() {
-	  if (sortedString.length > 0) {
-		var filename = "SortedText.txt";
-		var blob = new Blob([sortedString], { type: 'text/plain' });
-		var url = window.URL.createObjectURL(blob);
-  
-		var a = document.createElement('a');
-		a.href = url;
-		a.download = filename;
-		document.body.appendChild(a);
-		a.click();
-  
-		window.URL.revokeObjectURL(url);
-		document.body.removeChild(a);
-	  }
-	}
-  
-	function downloadPDF() {
-	  const doc = new jsPDF();
-	  doc.text(sortedString, 20, 20);
-	  doc.save('SortedText.pdf');
-	}
-  </script>
-  
-  <Intro heading={data.meta.title} description={data.meta.description} />
-  
-  <section class="bg-white dark:bg-gray-900">
-	<div class="py-8 px-4 mx-auto max-w-screen-xl lg:px-12">
-	  <div class="card p-8 relative items-center mx-auto max-w-screen-xl overflow-hidden rounded-lg">
-		<div class="rounded-lg overflow-hidden bg-gray-50 border border-gray-300" id="selectorBox">
-		  <select bind:value={sortingOption} class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500">
-			<option value="Select">Select</option>
-			<option value="alphabetical">Sort Alphabetically</option>
-			<option value="length">Sort by Length</option>
-			<option value="reverse">Reverse</option>
-			<option value="random">Random Sort</option>
-		  </select>
-		</div>
-		<div class="mt-3 gap-2 items-center mx-auto max-w-screen-xl lg:grid lg:grid-cols-2 overflow-hidden" id="boxarea">
-		  <div class="rounded-lg overflow-hidden bg-gray-50 border border-gray-300" id="tarea1">
-			<textarea placeholder="Enter Lines of text" id="textbox" class="block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500" rows="8" style="writing-mode: vertical-rl; white-space: nowrap;" on:input={handleInputChange}></textarea>
-		  </div>
-		  <div class="rounded-lg overflow-hidden bg-gray-50 border border-gray-300" id="tarea2">
-			<textarea placeholder="Output" id="textbox" rows="8" class="block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500" bind:value={sortedString} style="writing-mode: vertical-rl; white-space: nowrap;"></textarea>
-		  </div>
-		</div>
-		<div id="buttonArea">
-		  <Button color="blue" on:click={copyText}>Copy</Button>
-		  <Button color="blue" on:click={downloadText}>Download as txt</Button>
-		  <Button color="blue" on:click={downloadPDF}>Download as pdf</Button>
-		</div>
-	  </div>
-	</div>
-  </section>
-  
-  <style>
-	#textbox {
-	  resize: none;
-	}
-  
-	#boxarea {
-	  margin-top: 20px;
-	  gap: 20px;
-	}
-  
-	#buttonArea {
-	  margin-top: 30px;
-	  display: flex;
-	  justify-content: center;
-	  align-items: center;
-	  gap: 30px;
-	}
-  
-	.card {
-	  box-shadow: rgba(0, 0, 0, 0.1) 0 0 0 2px;
-	}
-  
-	:is(.dark .card) {
-	  box-shadow: rgba(255, 255, 255, 0.5) 0 0 0 2px;
-	}
-  </style>
