+++ src/routes/(tools)/remove-line-break/+page.svelte
@@ -1,22 +1,23 @@
 <script lang="ts">
-
+    
     import { Button } from 'flowbite-svelte';
     import jsPDF from 'jspdf';
     import Intro from '$lib/Intro.svelte';
 
     export let data;
+ 
+    let inputText = '';
+    let textWithoutLineBreaks = '';
 
-    var output = 0;
-
-    function countWords(input) {
-        const words = input.target.value.split(" ").filter(word => word !== "");
-        output = words.length;
+    function removeLineBreaks() {
+        textWithoutLineBreaks = inputText.replace(/\n/g, ' ');
     }
 
+   
     function copyText() {
-        if (output > 0) {
+        if (inputText.length > 0) {
             var textarea = document.createElement("textarea");
-            textarea.value = output;
+            textarea.value = textWithoutLineBreaks; 
             document.body.appendChild(textarea);
             textarea.select();
             document.execCommand("copy");
@@ -25,9 +26,9 @@
     }
 
     function downloadText() {
-        if (output > 0) {
+        if (inputText.length > 0) {
             var filename = "devstar_output.txt";
-            var blob = new Blob([output], { type: 'text/plain' });
+            var blob = new Blob([textWithoutLineBreaks], { type: 'text/plain' });
             var url = window.URL.createObjectURL(blob);
             
             var a = document.createElement('a');
@@ -35,16 +36,17 @@
             a.download = filename;
             document.body.appendChild(a);
             a.click();
-            
             window.URL.revokeObjectURL(url);
             document.body.removeChild(a);
         }
     }
-  
+
     function downloadPDF() {
-        const doc = new jsPDF();
-        doc.text(output.toString(), 20, 20);
-        doc.save('devstar_output.pdf');
+        if (textWithoutLineBreaks.length > 0) {
+            const pdf = new jsPDF();
+            pdf.text(textWithoutLineBreaks, 10, 10); 
+            pdf.save("devstar_output.pdf"); 
+        }
     }
 
 </script>
@@ -52,19 +54,22 @@
 <Intro heading={data.meta.title} description={data.meta.description} />
 
 <section class="bg-white dark:bg-gray-900">
-    <div class="py-8 px-4 mx-auto max-w-screen-xl lg:px-12">
-        <div class="card p-8 relative items-center mx-auto max-w-screen-xl overflow-hidden rounded-lg">
+	<div class="py-8 px-4 mx-auto max-w-screen-xl lg:px-12">
+		<div class="card p-8 relative items-center mx-auto max-w-screen-xl overflow-hidden rounded-lg">
 
-            <div class="gap-4 items-center mx-auto max-w-screen-xl lg:grid lg:grid-cols-2 overflow-hidden">
+			<div class="gap-4 items-center mx-auto max-w-screen-xl lg:grid lg:grid-cols-2 overflow-hidden">
+
                 <div class="rounded-lg overflow-hidden bg-gray-50 border border-gray-300">
                     <textarea placeholder="Enter Text" rows="8" class="resize-none block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-                    on:input={countWords}/>
+                    bind:value={inputText}  on:input={removeLineBreaks}/>
                 </div>
+
                 <div class="rounded-lg overflow-hidden bg-gray-50 border border-gray-300">
-                    <textarea placeholder="Word Count" rows="8" class="resize-none block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-                    bind:value={output}/>
+                    <textarea placeholder="Result" rows="8" class="resize-none block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                    bind:value={textWithoutLineBreaks}/>
                 </div>
-            </div>
+
+	        </div>
 
 			<div class="items-center mx-auto max-w-screen-xl lg:grid lg:grid-cols-1 overflow-hidden">
 				<div class="mt-8 gap-4 items-center mx-auto max-w-screen-xl lg:grid lg:grid-cols-3 overflow-hidden">
@@ -73,9 +78,9 @@
 					<Button color="blue" on:click={downloadPDF}>Download as pdf</Button>
 				</div>	
 			</div>
-            
-        </div>
-    </div>
+
+		</div>
+	</div>
 </section>
 
 <style>
