+++ src/routes/(tools)/strip-html-tags/+page.svelte
@@ -25,7 +25,7 @@
 
 	function downloadText() {
 		if (output.length > 0) {
-			var filename = "DevStar.txt";
+			var filename = "devstar_output.txt";
 			var blob = new Blob([output], { type: 'text/plain' });
 			var url = window.URL.createObjectURL(blob);
 			
@@ -43,7 +43,7 @@
 	function downloadPDF() {
 		const doc = new jsPDF();
 		doc.text(output, 20, 20);
-		doc.save('DevStaroutput.pdf');
+		doc.save('devstar_output.pdf');
 	}
 
 </script>
@@ -53,24 +53,27 @@
 <section class="bg-white dark:bg-gray-900">
 	<div class="py-8 px-4 mx-auto max-w-screen-xl lg:px-12">
 		<div class="card p-8 relative items-center mx-auto max-w-screen-xl overflow-hidden rounded-lg">
-			<div class="mt-3 gap-2 items-center mx-auto max-w-screen-xl lg:grid lg:grid-cols-2 overflow-hidden" id="boxarea">
 
-				<div class="rounded-lg overflow-hidden bg-gray-50 border border-gray-300" id="tarea1">
-					<textarea placeholder="Enter Text" id="textbox" rows="8" class="block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+			<div class="gap-4 items-center mx-auto max-w-screen-xl lg:grid lg:grid-cols-2 overflow-hidden">
+
+				<div class="rounded-lg overflow-hidden bg-gray-50 border border-gray-300">
+					<textarea placeholder="Enter Text" rows="8" class="resize-none block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
 					on:input={generateoutput}/>
 				</div>
 
-				<div class="rounded-lg overflow-hidden bg-gray-50 border border-gray-300" id="tarea2">
-					<textarea placeholder="Result" id="textbox" rows="8" class="block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+				<div class="rounded-lg overflow-hidden bg-gray-50 border border-gray-300">
+					<textarea placeholder="Result" rows="8" class="resize-none block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
 					bind:value={output}/>
 				</div>
 
 			</div>
 
-			<div id="buttonArea">
-				<Button color="blue" on:click={copyText}>Copy</Button>
-				<Button color="blue" on:click={downloadText}>Download as txt</Button>
-				<Button color="blue" on:click={downloadPDF}>Download as pdf</Button>
+			<div class="items-center mx-auto max-w-screen-xl lg:grid lg:grid-cols-1 overflow-hidden">
+				<div class="mt-8 gap-4 items-center mx-auto max-w-screen-xl lg:grid lg:grid-cols-3 overflow-hidden">
+					<Button color="blue" on:click={copyText}>Copy</Button>
+					<Button color="blue" on:click={downloadText}>Download as txt</Button>
+					<Button color="blue" on:click={downloadPDF}>Download as pdf</Button>
+				</div>	
 			</div>
 
 		</div>
@@ -78,22 +81,6 @@
 </section>
 
 <style>	
-
-	#textbox{
-		resize: none;
-	}
-
-	#boxarea{
-		gap:30px;
-	}
-
-	#buttonArea {
-		margin-top: 30px;
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		gap: 30px; /* Adjust the gap between buttons as needed */
-	}
 
 	.card {
 		box-shadow: rgba(0, 0, 0, 0.1) 0 0 0 2px;
